backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
jinja2==3.1.6
python-docx==1.1.2
qrcode[pil]==7.4.2
markdown==3.7
email-validator==2.2.0
itsdangerous==2.2.0
